Add compiler script of torch model to torchscript model (#639)

+++ model_compiler/tests/model_compiler/compilers/test_torch_model_file_to_onnx_model.py
@@ -18,27 +18,27 @@ class ConfigTestCase(TestCase):
     def test_from_json(self):
         self.assertEqual(Config.from_json({'input_names': ['input'],
                                            'input_shapes': [1, 2, 3, 4],
-                                           'data_type': 'FLOAT16',
+                                           'data_type': 'FLOAT',
                                            'max_batch_size': 1,
                                            'input_formats': ['channels_first']
                                            }),
                          Config(input_names=['input'],
                                 input_shapes=[[1, 2, 3, 4]],
                                 max_batch_size=1,
-                                data_type=torch.float16,
+                                data_type=torch.float,      # pylint: disable=no-member
                                 input_formats=[DataFormat.CHANNELS_FIRST]))
 
     def test_from_env(self):
         self.assertEqual(Config.from_env({'INPUT_NAMES': 'input',
                                           'INPUT_SHAPES': '[1, 2, 3, 4]',
                                           'MAX_BATCH_SIZE': '1',
-                                          'DATA_TYPE': 'FLOAT16',
+                                          'DATA_TYPE': 'FLOAT',
                                           'INPUT_FORMATS': 'channels_last'
                                           }),
                          Config(input_names=['input'],
                                 input_shapes=[[1, 2, 3, 4]],
                                 max_batch_size=1,
-                                data_type=torch.float16,
+                                data_type=torch.float,      # pylint: disable=no-member
                                 input_formats=[DataFormat.CHANNELS_LAST]))
